Remove digit shift from VSR

+++ programs/helium-sub-daos/src/utils.rs
@@ -165,14 +165,14 @@ pub trait PrecisePosition {
   fn voting_power_precise_locked_precise(
     &self,
     curr_ts: i64,
-    max_locked_vote_weight: u64,
+    max_locked_vote_weight: u128,
     lockup_saturation_secs: u64,
   ) -> Result<u128>;
 
   fn voting_power_precise_cliff_precise(
     &self,
     curr_ts: i64,
-    max_locked_vote_weight: u64,
+    max_locked_vote_weight: u128,
     lockup_saturation_secs: u64,
   ) -> Result<u128>;
 }
@@ -213,7 +213,7 @@ impl PrecisePosition for PositionV0 {
   fn voting_power_precise_locked_precise(
     &self,
     curr_ts: i64,
-    max_locked_vote_weight: u64,
+    max_locked_vote_weight: u128,
     lockup_saturation_secs: u64,
   ) -> Result<u128> {
     if self.lockup.expired(curr_ts) || (max_locked_vote_weight == 0) {
@@ -238,7 +238,7 @@ impl PrecisePosition for PositionV0 {
   fn voting_power_precise_cliff_precise(
     &self,
     curr_ts: i64,
-    max_locked_vote_weight: u64,
+    max_locked_vote_weight: u128,
     lockup_saturation_secs: u64,
   ) -> Result<u128> {
     let remaining = min(self.lockup.seconds_left(curr_ts), lockup_saturation_secs);
